# [WIP] Optimize RNN performance on CPU by fusing elemwise operations