This PR aims at improving LSTM and GRU performance by fusing elemwise operations in LSTMCell and GRUCell. Benefits:
1. save memory bandwidth
2. reduce autograd engine deduce

At functional level, this PR applies to all RNN features:
1. no special memory format for RNN weights, completely aligns with current weight format
2. inference and training
3. supports double backward
4. co-works with precompute of `input_gates` (time step fusion on input of weight_ih)
5. supports both `PackedSequence` and `PaddedSequence`
6. supports batch_first, dropout, multi layer, bidirectional

Performance:
Inference on single socket of Xeon Skylake 6148, **1.62x** improvement:
```bash
before:
lstm: layers=1, dir=1, batch_size=1, time_step=15, input_size=250, hidden_size=200: time: 0.73 ms
after:
lstm: layers=1, dir=1, batch_size=1, time_step=15, input_size=250, hidden_size=200: time: 0.45 ms
```
Inference on single core of Xeon Skylake 6148, **1.38x** improvement:
```bash
before:
lstm: layers=1, dir=1, batch_size=1, time_step=15, input_size=250, hidden_size=200: time: 0.77 ms
after:
lstm: layers=1, dir=1, batch_size=1, time_step=15, input_size=250, hidden_size=200: time: 0.56 ms
```
Training on single socket of Xeon Skylake 6148, **3.44x** improvement
```bash
before:
lstm: layers=1, dir=1, batch_size=128, time_step=50, input_size=1024, hidden_size=1024: time: 891.87 ms
after:
lstm: layers=1, dir=1, batch_size=128, time_step=50, input_size=1024, hidden_size=1024: time: 258.98 ms
```
All test results have config jemalloc, otherwise the "before" CPU performance would be even worse.

Benchmark script placed at [repo](https://github.com/mingfeima/op_bench-py), `run.sh rnn.py`. For inference, pick size according to suggestion; For training, pick size from RNN-translation of MLPerf.