@glaringlee 2020-08-03
@mingfeima 
It seems your code breaks our asan built (address sanitizer) tests in the following: (potential memory leak due to vectorization)
ci/circleci: pytorch_linux_xenial_py3_clang5_asan_test1 
ci/circleci: pytorch_linux_xenial_py3_clang5_asan_test2

They all call nn.LSTM.

One of failure is this:
https://app.circleci.com/pipelines/github/pytorch/pytorch/194791/workflows/41dd2318-66e5-416b-8f87-e9fb6e8151e5/jobs/6428572
The code is here:
https://github.com/pytorch/pytorch/blob/ca1b8ebbcb8f959c4044d5832acb34f1d9da179c/test/jit/test_models.py#L402

Another failure is this:
https://app.circleci.com/pipelines/github/pytorch/pytorch/194791/workflows/41dd2318-66e5-416b-8f87-e9fb6e8151e5/jobs/6428574
The code is here:
https://github.com/pytorch/pytorch/blob/ca1b8ebbcb8f959c4044d5832acb34f1d9da179c/test/test_nn.py#L11089

Can you take check your code?

Before doing any update, please rebase your code after the pytorch build status in the following link goes back to all green.
https://ezyang.github.io/pytorch-ci-hud/build2/pytorch-master
If you code can pass all the test, that's fine then, otherwise, there might be something wrong within your code.