@mingfeima 2020-08-28
@glaringlee sorry for the delay on this patch, i have been occupied by some other high priority jobs. 
The asan test errors shall be gone after replacing remainder vectorization math with scalar math. Though i thought this might be a false alarm, still it is better to have it fixed.
I am now considering to unify func signature from `_thnn_fused_lstm_cell` of both CPU and CUDA backend, thus the code would be much simplified.
Also I am integrating onednn/mkldnn rnn APIs into pytorch, will compare onednn/mkldnn with this thnn implementation from feature wise and performance wise.