@glaringlee 2020-09-02
@mingfeima been busy recently, I went through the modified part, it looks good to me.
Please fix the lint error in your next update, thanks a lot.