@mingfeima 2020-12-09
@glaringlee Ohhh... sorry for the delay. My github notifications are stupidly blocked by outlook.

About this patch, I was not so sure whether I want to continue to push it. 
Since the performance improvement is mediocre and too much change in RNN.cpp (due to the double backward support), making the rnn code a little bit too complexed.

Too achieve optimal performance, need to make sure the weight is consistently preserved in cache during the iteration of time sequence. Let's close this one for now.